予約の本人確認を追加

+++ app/Http/Controllers/ReservationCancelController.php
@@ -8,6 +8,9 @@ class ReservationCancelController extends Controller
 {
     public function destroy(Reservation $reservation)
     {
+        if ($reservation->user_id !== auth()->id()) {
+            abort(403);
+        }
         $reservation->delete();
 
         return redirect('/mypage');
